Changed all (?) occurences where validation macros (IS_xxx) had
* include/libxml/parserInternals.h HTMLparser.c HTMLtree.c
  SAX2.c catalog.c debugXML.c entities.c parser.c relaxng.c
  testSAX.c tree.c valid.c xmlschemas.c xmlschemastypes.c
  xpath.c: Changed all (?) occurences where validation macros
  (IS_xxx) had single-byte arguments to use IS_xxx_CH instead
  (e.g. IS_BLANK changed to IS_BLANK_CH).  This gets rid of
  many warning messages on certain platforms, and also high-
  lights places in the library which may need to be enhanced
  for proper UTF8 handling.

+++ testSAX.c
@@ -755,7 +755,7 @@ startElementNsDebug(void *ctx ATTRIBUTE_UNUSED,
 	    else
 		fprintf(stdout, ", %s='", attributes[i]);
 	    fprintf(stdout, "%.4s...', %d", attributes[i + 3],
-		    attributes[i + 4] - attributes[i + 3]);
+		    (int)(attributes[i + 4] - attributes[i + 3]));
 	}
     }
     fprintf(stdout, ")\n");
